+++ src/upmex/api/ecosystems.py
@@ -66,7 +66,9 @@ class EcosystemsAPI:
             PackageType.PYTHON_SDIST: "pypi.org",
             PackageType.NPM: "npmjs.org",
             PackageType.MAVEN: "repo.maven.apache.org",
-            PackageType.JAR: "repo.maven.apache.org"
+            PackageType.JAR: "repo.maven.apache.org",
+            PackageType.RUBY_GEM: "rubygems.org",
+            PackageType.RUST_CRATE: "crates.io"
         }
         return mapping.get(package_type)