+++ src/upmex/core/models.py
@@ -16,6 +16,8 @@ class PackageType(Enum):
     NPM = "npm"
     MAVEN = "maven"
     JAR = "jar"
+    RUBY_GEM = "ruby_gem"
+    RUST_CRATE = "rust_crate"
     GENERIC = "generic"
     UNKNOWN = "unknown"